refactor: completed successfully

+++ Generic-Collections-Datastructure/Models/Product.cs
@@ -1,6 +1,6 @@
 ﻿namespace Generic_Collections_Datastructure
 {
-   public abstract class Product
+    public abstract class Product
     {
         //  Product abstrakt class
         //- Id - qıraqdan set etmək olmayacaq yalnız get etmək olacaq və hər dəfə yeni bir Book obyekt yaradıldıqda avtomatik bir vahid artacaq.
@@ -15,7 +15,7 @@
         public string Name { get; set; }
         public double Price { get; set; }
         public static int Count { get; set; }
-        protected static  int TotalInCome { get; set; }
+        protected static double TotalInCome { get; set; }
 
 
         public Product()
@@ -27,7 +27,7 @@
         public abstract void Sell();
         public abstract void ShowInfo();
 
-        
+
 
     }
 }
